+++ frontend/src/components/inbox/fileUpload/context.tsx
@@ -1,5 +1,5 @@
-import { createContext, useContext, useState } from 'react';
 import React from 'react';
+import { createContext, useContext, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { Box } from '@mui/material';